Update tests, rename filters and selectors to be more meaningful
